Notes 19OCT2024
Noticed the R code did not conform to the RMD file. Fixed that for some of the lines. Not all though.

Also, was pointed out to me that gsub could be used more efficiently instead of making a function. So, I used regular expressions instead. Seems more reasonable really.

Cambridge animal report is a small effort that I forgot about, so I'll have to decided what to do with that later.